mpegts_pid_tracker: Fixes for crate usage
